+++ heftia-effects/src/Control/Monad/Hefty/Unlift.hs
@@ -16,12 +16,15 @@ where
 
 import Control.Monad.Hefty (Eff, interpretH, runEff, send0, type (~>))
 import Data.Effect.Unlift
+import UnliftIO (MonadUnliftIO)
+import UnliftIO qualified as IO
 
 runUnliftBase :: forall b. (Monad b) => Eff '[UnliftBase b] '[b] ~> b
 runUnliftBase =
     runEff . interpretH \(WithRunInBase f) ->
         send0 $ f runEff
 
-runUnliftIO :: Eff '[UnliftIO] '[IO] ~> IO
-runUnliftIO = runUnliftBase
-{-# INLINE runUnliftIO #-}
+runUnliftIO :: (MonadUnliftIO m) => Eff '[UnliftIO] '[m] ~> m
+runUnliftIO =
+    runEff . interpretH \(WithRunInBase f) ->
+        send0 $ IO.withRunInIO \run -> f $ run . runEff